Create some utility methods and move constants around
