+++ src/components/rooms/RoomCard.tsx
@@ -1,38 +1,19 @@
 import { Link } from 'react-router';
+import { Users } from 'lucide-react';
 import type { Room } from '@/types/room';
 import { ROUTES } from '@/constants/routes';
 import { ROOMS_PAGE_TEXTS } from '@/constants/room';
+import { createImageErrorHandler, createPlaceholderImage } from '@/utils/image';
 
 interface RoomCardProps {
   room: Room;
 }
 
-const UsersIcon = () => (
-  <svg
-    xmlns='http://www.w3.org/2000/svg'
-    viewBox='0 0 24 24'
-    fill='none'
-    stroke='currentColor'
-    strokeWidth='2'
-    strokeLinecap='round'
-    strokeLinejoin='round'
-    className='h-4 w-4'
-  >
-    <path d='M16 21v-2a4 4 0 0 0-4-4H6a4 4 0 0 0-4 4v2' />
-    <circle cx='9' cy='7' r='4' />
-    <path d='M22 21v-2a4 4 0 0 0-3-3.87' />
-    <path d='M16 3.13a4 4 0 0 1 0 7.75' />
-  </svg>
-);
-
 const RoomCard = ({ room }: RoomCardProps) => {
-  const handleImageError = (e: React.SyntheticEvent<HTMLImageElement>) => {
-    const target = e.currentTarget;
-    target.src = `https://via.placeholder.com/400/231a30/ffffff?text=${encodeURIComponent(room.name.substring(0, 2))}`;
-  };
+  const handleImageError = createImageErrorHandler(room.name.substring(0, 2), '400x400');
 
   const usersCount = room.members?.length || 0;
-  const defaultCoverImage = `https://via.placeholder.com/400/231a30/ffffff?text=${encodeURIComponent(room.name.substring(0, 2))}`;
+  const defaultCoverImage = createPlaceholderImage(room.name.substring(0, 2), '400x400');
 
   return (
     <Link
@@ -55,7 +36,7 @@ const RoomCard = ({ room }: RoomCardProps) => {
           <p className='text-text-secondary truncate text-sm'>No active music</p>
         )}
         <div className='text-text-secondary flex items-center gap-1.5 pt-1'>
-          <UsersIcon />
+          <Users className='h-4 w-4' />
           <span className='text-sm'>
             {usersCount} {ROOMS_PAGE_TEXTS.USERS_LABEL}
           </span>
