+++ src/pages/Room.tsx
@@ -1,16 +1,13 @@
 import { useState, useEffect } from 'react';
 import { useParams, useNavigate } from 'react-router';
-import Layout from '@/components/layout/Layout';
+import BaseLayout from '@/components/layout/BaseLayout';
 import SongSearch from '@/components/room/SongSearch';
 import SongPlayer from '@/components/room/SongPlayer';
 import SongQueueTabs from '@/components/room/SongQueueTabs';
 import type {
-  Song,
   QueueSong,
   SearchResult,
   PlayerState,
-  HistorySong,
-  RoomSession,
   ChatMessage,
   Room as RoomType,
 } from '@/types/room';
@@ -18,164 +15,14 @@ import { roomsService } from '@/services/rooms';
 import { spotifyService } from '@/services/spotify';
 import { useAuthStore } from '@/stores/auth';
 import { ROUTES } from '@/constants/routes';
-
-// Mock data for development
-const MOCK_CURRENT_SONG: Song = {
-  id: '1',
-  name: 'Blinding Lights',
-  artist: 'The Weeknd',
-  albumCover: 'https://images.unsplash.com/photo-1493225457124-a3eb161ffa5f?w=300&h=300&fit=crop',
-  duration: 200,
-};
-
-const MOCK_QUEUE: QueueSong[] = [
-  {
-    id: '2',
-    name: 'As It Was',
-    artist: 'Harry Styles',
-    albumCover: 'https://images.unsplash.com/photo-1511671782779-c97d3d27a1d4?w=300&h=300&fit=crop',
-    duration: 167,
-    likes: 12,
-    dislikes: 2,
-    addedBy: 'user1',
-  },
-  {
-    id: '3',
-    name: 'Save Your Tears',
-    artist: 'The Weeknd',
-    albumCover: 'https://images.unsplash.com/photo-1493225457124-a3eb161ffa5f?w=300&h=300&fit=crop',
-    duration: 215,
-    likes: 8,
-    dislikes: 1,
-    addedBy: 'user2',
-  },
-];
-
-const MOCK_RECENT_SONGS: HistorySong[] = [
-  {
-    id: '10',
-    name: 'Starboy',
-    artist: 'The Weeknd',
-    albumCover: 'https://images.unsplash.com/photo-1493225457124-a3eb161ffa5f?w=300&h=300&fit=crop',
-    duration: 230,
-    playedAt: new Date(Date.now() - 1000 * 60 * 5),
-    addedBy: 'user3',
-  },
-  {
-    id: '11',
-    name: 'Levitating',
-    artist: 'Dua Lipa',
-    albumCover: 'https://images.unsplash.com/photo-1511671782779-c97d3d27a1d4?w=300&h=300&fit=crop',
-    duration: 203,
-    playedAt: new Date(Date.now() - 1000 * 60 * 10),
-    addedBy: 'user1',
-  },
-];
-
-const MOCK_SESSIONS: RoomSession[] = [
-  {
-    id: 'session-1',
-    date: new Date(Date.now() - 1000 * 60 * 60 * 24),
-    songs: [
-      {
-        id: '20',
-        name: 'Shape of You',
-        artist: 'Ed Sheeran',
-        albumCover:
-          'https://images.unsplash.com/photo-1493225457124-a3eb161ffa5f?w=300&h=300&fit=crop',
-        duration: 234,
-        playedAt: new Date(Date.now() - 1000 * 60 * 60 * 24 - 1000 * 60 * 30),
-        addedBy: 'user1',
-      },
-      {
-        id: '21',
-        name: 'Uptown Funk',
-        artist: 'Bruno Mars',
-        albumCover:
-          'https://images.unsplash.com/photo-1511671782779-c97d3d27a1d4?w=300&h=300&fit=crop',
-        duration: 270,
-        playedAt: new Date(Date.now() - 1000 * 60 * 60 * 24 - 1000 * 60 * 25),
-        addedBy: 'user2',
-      },
-      {
-        id: '22',
-        name: "Can't Stop the Feeling",
-        artist: 'Justin Timberlake',
-        albumCover:
-          'https://images.unsplash.com/photo-1516280440614-37939bbacd81?w=300&h=300&fit=crop',
-        duration: 237,
-        playedAt: new Date(Date.now() - 1000 * 60 * 60 * 24 - 1000 * 60 * 20),
-        addedBy: 'user3',
-      },
-    ],
-  },
-  {
-    id: 'session-2',
-    date: new Date(Date.now() - 1000 * 60 * 60 * 48),
-    songs: [
-      {
-        id: '30',
-        name: 'Believer',
-        artist: 'Imagine Dragons',
-        albumCover:
-          'https://images.unsplash.com/photo-1514320291840-2e0a9bf2a9ae?w=300&h=300&fit=crop',
-        duration: 204,
-        playedAt: new Date(Date.now() - 1000 * 60 * 60 * 48 - 1000 * 60 * 15),
-        addedBy: 'user1',
-      },
-      {
-        id: '31',
-        name: 'Thunder',
-        artist: 'Imagine Dragons',
-        albumCover:
-          'https://images.unsplash.com/photo-1514320291840-2e0a9bf2a9ae?w=300&h=300&fit=crop',
-        duration: 187,
-        playedAt: new Date(Date.now() - 1000 * 60 * 60 * 48 - 1000 * 60 * 10),
-        addedBy: 'user2',
-      },
-    ],
-  },
-  {
-    id: 'session-3',
-    date: new Date(Date.now() - 1000 * 60 * 60 * 72),
-    songs: [
-      {
-        id: '40',
-        name: 'Photograph',
-        artist: 'Ed Sheeran',
-        albumCover:
-          'https://images.unsplash.com/photo-1493225457124-a3eb161ffa5f?w=300&h=300&fit=crop',
-        duration: 258,
-        playedAt: new Date(Date.now() - 1000 * 60 * 60 * 72 - 1000 * 60 * 10),
-        addedBy: 'user4',
-      },
-    ],
-  },
-];
-
-const MOCK_CHAT_MESSAGES: ChatMessage[] = [
-  {
-    id: 'msg-1',
-    userId: 'user1',
-    userName: 'Carlos',
-    message: 'What a great song! 🎵',
-    timestamp: new Date(Date.now() - 1000 * 60 * 10),
-  },
-  {
-    id: 'msg-2',
-    userId: 'user2',
-    userName: 'Maria',
-    message: 'Can someone play some rock?',
-    timestamp: new Date(Date.now() - 1000 * 60 * 8),
-  },
-  {
-    id: 'msg-3',
-    userId: 'user3',
-    userName: 'Pablo',
-    message: "I'll add one right now",
-    timestamp: new Date(Date.now() - 1000 * 60 * 5),
-  },
-];
+import {
+  MOCK_CURRENT_SONG,
+  MOCK_QUEUE,
+  MOCK_RECENT_SONGS,
+  MOCK_SESSIONS,
+  MOCK_CHAT_MESSAGES,
+} from '@/constants/mockData';
+import { checkIsHost } from '@/utils/room';
 
 const Room = () => {
   const { id: roomCode } = useParams<{ id: string }>();
@@ -186,17 +33,16 @@ const Room = () => {
   const [isLoading, setIsLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
   const [isJoining, setIsJoining] = useState(false);
-
   const [playerState, setPlayerState] = useState<PlayerState>({
     currentTime: 140,
     isPlaying: true,
     volume: 0.8,
   });
-
   const [queue, setQueue] = useState<QueueSong[]>(MOCK_QUEUE);
   const [searchResults, setSearchResults] = useState<SearchResult[]>([]);
   const [isSearching, setIsSearching] = useState(false);
   const [messages, setMessages] = useState<ChatMessage[]>(MOCK_CHAT_MESSAGES);
+  const isHost = checkIsHost(room, user);
 
   useEffect(() => {
     const loadRoomDetails = async () => {
@@ -273,10 +119,7 @@ const Room = () => {
 
     if (!roomCode || !room) return;
 
-    // Check if user is the host by comparing with host_id or host.spotify_id
-    const userIsHost = checkIsHost(room, user.id);
-
-    if (!userIsHost) {
+    if (!isHost) {
       setError('Only the host can close the room');
       return;
     }
@@ -294,21 +137,6 @@ const Room = () => {
       console.error('Error closing room:', err);
     }
   };
-
-  // Check if current user is the room host
-  const checkIsHost = (roomData: RoomType | null, userId?: string): boolean => {
-    if (!roomData || !userId) return false;
-
-    // First try with host_id (user UUID)
-    if (roomData.host_id === userId) return true;
-
-    // If host object exists, compare with spotify_id
-    if (roomData.host?.spotify_id === user?.spotify_id) return true;
-
-    return false;
-  };
-
-  const isHost = checkIsHost(room, user?.id);
 
   const handleTogglePlay = () => {
     setPlayerState((prev) => ({ ...prev, isPlaying: !prev.isPlaying }));
@@ -361,18 +189,18 @@ const Room = () => {
 
   if (isLoading) {
     return (
-      <Layout>
-        <div className='flex h-[calc(100vh-56.5px)] items-center justify-center'>
+      <BaseLayout>
+        <div className='flex h-[calc(100dvh-61px)] items-center justify-center'>
           <div className='text-text-secondary'>Loading room...</div>
         </div>
-      </Layout>
+      </BaseLayout>
     );
   }
 
   if (error && !room) {
     return (
-      <Layout>
-        <div className='flex h-[calc(100vh-56.5px)] flex-col items-center justify-center gap-4'>
+      <BaseLayout>
+        <div className='flex h-[calc(100dvh-61px)] flex-col items-center justify-center gap-4'>
           <div className='text-red-500'>{error}</div>
           <button
             onClick={() => navigate(ROUTES.ROOMS)}
@@ -381,23 +209,23 @@ const Room = () => {
             Back to rooms
           </button>
         </div>
-      </Layout>
+      </BaseLayout>
     );
   }
 
   if (isJoining) {
     return (
-      <Layout>
-        <div className='flex h-[calc(100vh-56.5px)] items-center justify-center'>
+      <BaseLayout>
+        <div className='flex h-[calc(100dvh-61px)] items-center justify-center'>
           <div className='text-text-secondary'>Joining the room...</div>
         </div>
-      </Layout>
+      </BaseLayout>
     );
   }
 
   return (
-    <Layout>
-      <div className='flex h-[calc(100vh-56.5px)] flex-col py-4'>
+    <BaseLayout>
+      <div className='flex h-[calc(100dvh-61px)] flex-col py-4'>
         {/* Room header with controls */}
         <div className='px-6 pb-4 md:px-12 xl:px-6'>
           <div className='flex items-center justify-between'>
@@ -466,7 +294,7 @@ const Room = () => {
           </div>
         </div>
       </div>
-    </Layout>
+    </BaseLayout>
   );
 };
