+++ src/utils/format.ts
@@ -1,0 +1,17 @@
+/**
+ * Formats seconds to mm:ss format
+ */
+export const formatTime = (seconds: number): string => {
+  const mins = Math.floor(seconds / 60);
+  const secs = Math.floor(seconds % 60);
+  return `${mins}:${secs.toString().padStart(2, '0')}`;
+};
+
+/**
+ * Formats a Date object to time string (HH:MM AM/PM)
+ */
+export const formatMessageTime = (date: Date): string =>
+  date.toLocaleTimeString('en-US', {
+    hour: '2-digit',
+    minute: '2-digit',
+  });
